requirements.txt
alembic==1.7.4
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
passlib==1.7.4
PyJWT==2.3.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
speaklater==1.3
SQLAlchemy==1.4.26
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0
marshmallow==3.14.0
psycopg2-binary==2.9.1
flask-redis==0.4.0
flasgger==0.9.5
grpcio-tools==1.42.0
rauth==0.7.2
python-decouple

